requirements.txt
# Install Python 3.10 before proceeding.

# Run the following before compiling Python 3.10 from source:
# sudo apt install liblzma-dev

# Run the following (before installing the pip package 'opencv-python'):
# sudo apt install libgtk2.0-dev pkg-config

# Run the following (before installing the pip package 'pyaudio'):
# sudo apt install portaudio19-dev

# Make sure Ollama (with the Phi-4 model) and ffmpeg are installed.

# sudo apt install ffmpeg libavcodec-dev libavformat-dev libavutil-dev libavdevice-dev libavfilter-dev libswscale-dev libswresample-dev pkg-config
# sudo apt install liblzma-dev lzma-dev

# To install the below python packages, create a virtual envifronment (venv), activate it, then place this file in the project directory and run 'pip install -r requirements.txt'
# Scroll down this file when the command finishes, you will have some dependency conflicts to take care o.

numpy>=2
absl-py==2.3.1
attrs==25.3.0

audioread==3.0.1
beartype==0.18.5
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
chatterbox-tts==0.1.2
click==8.2.1
coloredlogs==15.0.1
conformer==0.3.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.3
decorator==5.2.1
diffq==0.2.4
diffusers==0.29.0
distlib==0.4.0
einops==0.8.1
filelock==3.18.0
flash==1.0.3
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
future==1.0.0
hf-xet==1.1.7
huggingface-hub==0.34.4
humanfriendly==10.0
identify==2.6.13
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iso8601==2.1.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
ml_collections==1.1.0
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.2

nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-mediapipe==0.10.21cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.18.0
onnx-weekly==1.19.0.dev20250726
onnx2torch-py313==1.6.0
onnxruntime-gpu==1.22.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.3.0
protobuf==4.25.8
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
pygame==2.6.1
PyMatting==1.1.14
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
rembg==2.0.67
requests==2.32.4
resampy==0.4.3
resemble-perth==1.0.1
rotary-embedding-torch==0.6.5
rpds-py==0.27.0
s3tokenizer==0.2.0
safetensors==0.5.3
samplerate==0.1.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.1
serial==0.0.97
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SpeechRecognition==3.14.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.20.3



tqdm==4.67.1
transformers==4.46.3
triton==3.4.0
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.34.0
Werkzeug==3.1.3
zipp==3.23.0
onnxruntime-gpu
ollama

# Install remaining packages individually if meaningless dependency error prevents installation

# pip install mediapipe

# pip install lipsync
# pip uninstall av
# pip install av==15.0.0

# pip uninstall numpy
# pip install 'numpy>=2'

# pip uninstall torch torchvision torchaudio
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

# pip install audio-separator
# pip install rembg

# pip install onnxruntime-gpu

# After succesfully installing the above, pip should report the following dependency issues (can be checked with 'pip check'), which can be treated as normal:
# /////////////
# mediapipe 0.10.21 has requirement numpy<2, but you have numpy 2.2.6.
# lipsync 0.1.0 has requirement av<15.0.0,>=14.0.1, but you have av 15.0.0.
# chatterbox-tts 0.1.2 has requirement librosa==0.11.0, but you have librosa 0.10.2.post1.
# chatterbox-tts 0.1.2 has requirement torch==2.6.0, but you have torch 2.8.0+cu128.
# chatterbox-tts 0.1.2 has requirement torchaudio==2.6.0, but you have torchaudio 2.8.0+cu128.
# //////////////
# Any dependency issues aside from the ones above should be resolved, unless the programs works as expected anyway.